parsers/grammar: replace lark with tatsu
This is a huge change. It was motivated by the inability to concisely
describe the grammar (which is unambigous) in an unambigous form in CNF.
In the previous grammar, I had to be very careful not to accidentally
parse everything as a bunch of WORDs, but then all special forms had to
be listed one-by-one (see commit history for this, ugh).

PEG is much better, because the "|" operator is inherently ordered, and
it's very natural to describe "try to find X, or else fall back to WORD".
This made the grammar simple again, and unsurprisingly, the parser is also
way faster.

Tatsu is also way better in the AST front, because now the parsed parts of
the rules are simply object properties, I no longer need to bother with
"children[4]" kind of nonsense.

Testing for result stability is also no longer needed, because the grammar
is inherently unambigous.

Unfortunately, this parser actually generates python code, which we do
the first time the module is loaded. I know it will cause problems with
packaging later on, but right now, I have no better idea for my use-cases.
The nice thing though is that after the first run, the IDE has autocomplete
for all Model classes.