parser/grammar: rudimentary Subtitle-enabled parsing
This is heavily Btk. focused (see the tests), as these were all the
BlockAmendments related to Btk. that we were unable to properly parse.

Now we can do that. The metadata semantics may or may not break in the
future, especially if I realize that some more data is necessary to
actually apply amendments.