parsing: Slice MK into Acts.
In the python version, slicing into sections and then slicing the Acts
section into Acts was two distinct steps. The section slicer was also way
more flexible, but it was never used for any other section type, so in this
implementation it was replaced with an extremely simple algorithm.

No tests yet. It never had a standalone test in python, but there was
really no need for it either once it was written.