grammar: always try to collate reference parts, even if suffix is present
See the testdata for 'evil' cases. In the python version, these were only
parsed this way if they were part of a bigger structure (like repeal).
There's no reason to do this, and there was even a TODO in the code
mentioning that this may be a deficiency in the grammar.

Now that I'm older and wiser, I think it was in fact a deficiency and not
whatever that weird argument that was written there.