structure: move sae.semantic_info out of the Option<>
It was meant to signify if there was any semantic processing done or not,
but it should always be obvious from the context, and is only good for a
bunch of assertions which complicate the code, and bloat in serialization.

We _may_ need to introduce it as Option<Box<...>> for memory usage reason
though.