project: upgrade serde_yaml to 0.9, introduce singleton_yaml module
Originally in 0.8, the yaml module serialized enum values the same way as
it's done in JSON: to a dict with a single field.

0.9 changed the backend, and also introduced the tagged serialization of
enums. Unfortunately this is not compatible with the original
serialization, and also breaks the json<->yaml bijection.
There's also a limtiation in serde regarding tagged and untagged enums:
https://github.com/dtolnay/serde-yaml/issues/318

So we are staying on the singleton serialization, and I made a small
wrapper module for convenience.