fixups: add fixups for Btk-related acts
Support for act-like-decree references and abbreviations is not
implemented, so these were "disabled" in a rather crude way. Please delete
them when support lands.

For some reason a lot of uppercase chars are extracted as lowercase, this
should be fixed in the PDF parser some day.

In 2017/CXLIV the "f )" is also a PDF extraction error. If this problem
occurs anywhere else, it should be fixed in the extractor.

The whole "Alapítvány tv." thing in 2021/CI is a mess. Maybe abbreviations
should be handled completely differently. One thing is for sure: this
should not be solved in the grammar itself.